requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
bitarray==3.5.1
blinker==1.9.0
cachetools==4.2.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cryptg==0.5.0.post0
cryptography==45.0.5
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
filelock==3.18.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
hexbytes==1.3.1
httpcore==0.16.3
httpx==0.23.3
idna==3.10
Jinja2==3.1.6
jsonalias==0.1.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
lxml==6.0.0
MarkupSafe==3.0.2
multidict==6.6.3
narwhals==1.47.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parsimonious==0.10.0
pillow==11.3.0
propcache==0.3.2
protobuf==6.31.1
pyaes==1.6.1
pyarrow==20.0.0
pyasn1==0.6.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyOpenSSL==25.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==13.15
pytz==2025.2
pyunormalize==16.0.0
pywin32==311
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3986==1.5.0
rlp==4.1.0
rpds-py==0.26.0
rsa==4.9.1
schedule==1.2.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snscrape==0.7.0.20230622
solana==0.31.0
solders==0.19.0
soupsieve==2.7
streamlit==1.46.1
Telethon==1.40.0
tenacity==9.1.2
toml==0.10.2
toolz==1.0.0
tornado==6.1
types-cachetools==4.2.10
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
watchdog==6.0.0
web3==7.12.0
websockets==11.0.3
yarl==1.20.1
psutil
streamlit-js
